factorisation algo morpions dans un fichier general
Je n'ai pas reussi a factoriser la méthode de class nb_vertices_created proprement
